Add `call` to assert that an arbitrary closure won't panic
This function calls the given closure, asserting that there's no
possibility of panicking.  If the compiler can't prove this, the code
will be left with a `dont_panic!` linking error.

With the `panic` feature turned on, this function just calls the closure
directly, letting it panic or not on its own.